Fix ERC -> EIP Naming
I mistakenly assumed that ERC was an EIP that was final. This is not
the case. ERCs are for proposing standards related to the Ethereum
ecosystem (such as smart contract standards) while EIPs are changes to
the Ethereum protocol. These transactions types are all changes to the
protocol and are therefore EIPs.

Thanks @bh2smith for raising this!